Utilize new module structure for the API
